refactoring , honestly its been a bit everywhere, join now can join mutliple channels in 1 string , prefix function added to meassge builder

+++ v1.2/includes/Channel.hpp
@@ -43,7 +43,7 @@ struct WeakPtrCompare {
  class Channel {
 	private:
 		int _ClientLimit = 0;
-		int _clientCount = 0;
+		unsigned long _clientCount = 0;
 
 		unsigned long _ulimit;
 
@@ -63,7 +63,7 @@ struct WeakPtrCompare {
 
 		const std::string& getName() const;
 		const std::string& getTopic() const;
-		const int& getClientCount() {return _clientCount;};
+		const unsigned long& getClientCount() {return _clientCount;};
 		std::vector<int> getAllfds();
 		const std::string getAllNicknames();
 		std::weak_ptr<Client> getWeakPtrByNickname(const std::string& nickname);
